fix(daemon): preserve quality gates for peer discovery
